test: Implement integrationTest for performanceEntity and refactor unit test
