--- src/main/kotlin/com/group4/ticketingservice/config/MapperConfig.kt
@@ -1,15 +1,0 @@
-package com.group4.ticketingservice.config
-
-import org.modelmapper.ModelMapper
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class MapperConfig {
-    @Bean
-    fun modelMapper(): ModelMapper {
-        val modelMapper = ModelMapper()
-        modelMapper.configuration.isFieldMatchingEnabled = true
-        return ModelMapper()
-    }
-}
